@@ -3,26 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.2",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.20.4",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
-    "@types/react-router-dom": "^5.1.7",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/material": "^5.7.0",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.33",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@types/react-router-dom": "^5.3.3",
     "atoms": "^0.2.0",
     "jotai": "^0.16.3",
-    "mqtt": "^4.2.6",
-    "react": "^17.0.1",
+    "mqtt": "^4.3.7",
+    "react": "^18.1.0",
     "react-circular-progressbar": "^2.0.4",
-    "react-dom": "^17.0.1",
-    "react-modal": "^3.14.3",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.2.1",
-    "typescript": "^4.1.5",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.6.4",
     "web-vitals": "^1.1.0"
   },
   "scripts": {
@@ -51,7 +48,6 @@
   },
   "homepage": "/van/",
   "devDependencies": {
-    "@types/react-modal": "^3.12.1",
-    "@types/styled-components": "^5.1.7"
+    "@types/react-modal": "^3.13.1"
   }
 }
