@@ -1,25 +1,27 @@
-import { Typography } from "@material-ui/core";
-import React from "react";
+import styled from "@emotion/styled";
+import { Typography } from "@mui/material";
 import "react-circular-progressbar/dist/styles.css";
-import styled from "styled-components";
 import MqttLevel from "../mqttDevices/MqttLevel/MqttLevel";
 import MqttWeight from "../mqttDevices/MqttWeight/MqttWeight";
 
-const Headline = styled.div`
-  padding-top: 5vh;
-  display: flex;
-  justify-content: center;
-  color: ${p => p.theme.globalPage.color};
-`;
+const Headline = styled("div")({
+  display: "flex"
+})
+  // padding-top: 5vh;
+  // display: flex;
+  // justify-content: center;
+  // color: ${p => p.theme.globalPage.color};
 
-const Wrapper = styled.div`
-  padding: 40px 100px;
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-  align-items: center;
-  color: ${p => p.theme.globalPage.color};
-`;
+
+const Wrapper = styled("div")({
+  padding: "40px 100px",
+  display: "flex"
+})
+  
+  // justify-content: center;
+  // flex-wrap: wrap;
+  // align-items: center;
+  // color: ${p => p.theme.globalPage.color};
 
 export default function Water() {
   return (
