@@ -1,13 +1,13 @@
-import styled from "styled-components";
+import { styled } from "@mui/material";
 
-const BottomNav = styled.div`
-  width: 100%;
-  height: 50px;
-  position: sticky;
-  top: 0;
-  background-color: black;
-`;
-
+const BottomNav = styled("div")({
+  width: "100%",
+  height: "50px",
+  position: "sticky",
+  top: 0,
+  backgroundColor: "black"
+})
+  
 export default function Footer() {
   return <BottomNav></BottomNav>;
 }
