@@ -1,61 +1,57 @@
-import React from "react";
+import { styled } from "@mui/material";
 import { Link } from "react-router-dom";
-import styled from "styled-components";
 import DarkmodeSwitch from "./components/DarkmodeSwitch";
 import { ConnectionStatus, useMqttContext } from "./context/MqttContext";
 import { useSettingsContext } from "./context/SettingContext";
 import { dark, light } from "./styles/themes";
 
-const Menu = styled.div`
-  z-index: 100;
-  width: 100%;
-  position: sticky;
-  top: 0;
-  background-color: ${p => p.theme.secondary};
-  box-shadow: 0 10px 6px -6px ${p => p.theme.shadowColor};
-`;
 
-const Nav = styled.nav`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  padding: 16px 24px;
-`;
+const Menu = styled("div")(({theme}) => ({
+  width: "100%",
+  zIndex: 100,
+  position: "sticky",
+  top: "0",
+  color: theme.palette.primary.main,
+  boxShadow: theme.shadows[1]
+}))
 
-const FlowContainer = styled.div`
-  display: flex;
-  align-items: center;
-`;
 
-const Status = styled.div<{ status: ConnectionStatus }>`
-  height: 20px;
-  width: 20px;
-  margin-right: 10px;
-  border-radius: 50%;
-  border: 1px solid black;
-  background-color: ${p =>
-    p.status === ConnectionStatus.Connectet
-      ? "green"
-      : p.status === ConnectionStatus.Reconnect
-      ? "yellow"
-      : "red"};
-`;
+const Nav = styled("nav")({
+  display: "flex",
+  flexDirection: "row",
+  justifyContent: "space-between",
+  padding: "16px 24px"
+})
 
-const LinkButton = styled(Link)`
-  padding: 6px;
-  text-decoration: none;
-  color: ${p => p.theme.globalPage.color};
-  font-size: 16px;
+const FlowContainer = styled("div")({
+  display: "flex",
+  alignItems: "center"
+})
 
-  a {
-    text-decoration: none;
+const Status = styled("div")<{status: ConnectionStatus}>(({status}) => ({
+  height: "20px",
+  width: "20px",
+  marginRight: "10px",
+  borderRadius: "50%",
+  border: "1px solid black",
+  backgroundColor: status === ConnectionStatus.Connectet ? "green" : "red"
+}))
+  
+
+const LinkButton = styled(Link)({
+  padding: "6px",
+  textDecoration: "none",
+  fontSize: "16px",
+  
+  "a": {
+    textDecoration: "none",
   }
-`;
+})
 
-const Titel = styled.span`
-  font-size: 18px;
-  font-weight: bold;
-`;
+const Titel = styled("span")({
+  fontSize: "18px",
+  fontWeight: "bold"
+})
 
 export default function Header() {
   const { theme, setTheme } = useSettingsContext();
@@ -69,11 +65,11 @@ export default function Header() {
         </LinkButton>
 
         <FlowContainer>
-          <Status status={connectStatus} />
+          <Status status={connectStatus}/>
 
           <DarkmodeSwitch
             id={"ID"}
-            checked={theme?.themeName === "dark"}
+            checked={theme?.palette.mode === "dark"}
             onChange={e => setTheme(e.target.checked ? dark : light)}
           />
         </FlowContainer>
