@@ -1,16 +1,14 @@
+import { styled, ThemeProvider } from "@mui/material";
 import React, { ReactNode } from "react";
 import { BrowserRouter as Router } from "react-router-dom";
-import styled, { ThemeProvider } from "styled-components";
 import { useSettingsContext } from "./context/SettingContext";
 import Header from "./Header";
 import { light } from "./styles/themes";
 
-const Content = styled.div``;
-
-const Main = styled.main`
-  background-color: ${p => p.theme.globalPage.background};
-  min-height: 100vh;
-`;
+const Main = styled("main")({
+  minHeight: "100vh"
+})
+  // background-color: ${p => p.theme.globalPage.background};
 
 type Props = {
   children: ReactNode;
@@ -24,7 +22,7 @@ export default function Skeleton(props: Props) {
         <ThemeProvider theme={theme ? theme : light}>
           <Main>
             <Header></Header>
-            <Content>{props.children}</Content>
+            <div>{props.children}</div>
           </Main>
         </ThemeProvider>
       </Router>
