@@ -1,6 +1,6 @@
+import { styled } from "@mui/material";
 import React, { useEffect, useState } from "react";
 
-import styled from "styled-components";
 import { DashButton } from "../../components/dashContainer";
 import { useMqttContext } from "../../context/MqttContext";
 import { useSettingsContext } from "../../context/SettingContext";
@@ -34,28 +34,33 @@ type Props = {
   name: string;
 };
 
-const LevelContainer = styled.div`
-  height: 180px;
-  width: 50px;
-  display: flex;
-  flex-direction: column-reverse;
-  border-radius: 10px;
-  background-color: #777;
-`;
+const LevelContainer = styled("div")({
+  height: "180px",
+  width: "50px",
+})
+  // height: 180px;
+  // width: 50px;
+  // display: flex;
+  // flex-direction: column-reverse;
+  // border-radius: 10px;
+  // background-color: #777;
 
-const Level = styled.div<{ level: number }>`
-  bottom: 0;
-  left: 0;
-  background-color: #c4c10d;
-  border-radius: 10px;
-  height: ${p => (p.level > 10 ? p.level : 10)}%;
+const Level = styled("div")({
+  bottom: 0
+})
+// <{ level: number }>`
+//   bottom: 0;
+//   left: 0;
+//   background-color: #c4c10d;
+//   border-radius: 10px;
+//   height: ${p => (p.level > 10 ? p.level : 10)}%;
 
-  > div {
-    font-size: 15px;
-    font-weight: bold;
-    margin-top: ${p => (p.level > 13 ? 10 : 0)}px;
-  }
-`;
+//   > div {
+//     font-size: 15px;
+//     font-weight: bold;
+//     margin-top: ${p => (p.level > 13 ? 10 : 0)}px;
+//   }
+// `;
 
 export default function MqttWeight({ mqttTopic, name }: Props) {
   const [value, setValue] = useState<number>(0);
@@ -104,7 +109,7 @@ export default function MqttWeight({ mqttTopic, name }: Props) {
     <DashButton>
       <span>{name}</span>
       <LevelContainer>
-        <Level level={value}>
+        <Level>
           <div>{value}%</div>
         </Level>
       </LevelContainer>
